Variety of updates.
Added user settings file for configuring variables. This file is copied to .launcher.settings.sh.

This is done so that when a user pulls the repo, the default user configuration file can be updated, without destroying old user settings.

Settings can then be managed by the user. Instructions are displayed telling the user how to do this.

.update.sh tracks if the configuration has been updated or not.

Added .gitignore so that specific settings files are not set to the repo.

Updated readme.