Sub directory automation.
Sub directories will now automatically be searched.

Default searches are for /app, /sites, /wp-content.

If not found, the .docroots file is checked per line for other knowns webroots.

If not found there, the user will be asked for the framework webroot. The response is then written to the .docroots file, the provided sub-directory is checked, if exists, is opened. If it doesn’t exist a message is displayed.

In the future when a site is opened, the new entry within the .docroots file will be opened automatically.