Updated readme functionality walkthrough.
This is especially important now, as since functionality is now modularized it isn’t clear in which order code is executed.

You can of course just look at the launcher file, but I’ve changed the functionality walkthrough section to be defined by file and order each file is executed.

This may not be exactly correct, as some definitions do not matter, such as help or settings migrations.

The general goal of clarity is still achieved.