Fixed terminus issue.
No matter what is passed, dev is ran first and should not attempt to deploy.

Code was attempting to run a deploy on dev when ‘test’ or ‘live’ were passed.

Checked instead that on the first loop, always run the dev version. All other loops run the deploy step version. This fixes the error.