Attempting…
Trying to use terminus env:wake to test if an environment database exists or not.

The test itself works, but need to test against the console ouput messages for success or failure, not against $?.

Haven’t been able to figure out how to test against this output.

a;lsdkfja;lskdjf;alksdjf